chore: import netipx from rbmk-project/x
I cannot actually add netipx to x because the x package depends
on dnscore, so dnscore cannot depend on x.

While this is annoying, it's also true that netipx is good enough
to qualify for being inside the common package collection.

Therefore, just move it here and I'll update `x` accordingly.