BREAKING CHANGE(cliutils): inject command Environment
This diff modifies cliutils.Command to inject an Environment
interface for both Help and Main, which provides the stdout and
the stderr via methods. In turn, this change allows for a fine
grained customisation of `rbmk` commands stdout and stderr,
with immediate benefits for testing code, where we will be able
to drop the current singleton based approach.