feat(cliutils): add Stdin to Environment
This diff adds Stdin to the Environment type, thus allowing
to replace also the stdin when executing commands.

There is an `rbmk sh` command in `rbmk` and currently it
unconditionally uses the os.Stdin, which does not seem about
right and may backfire if we start to use commands in
non-conventional ways (e.g., executing them by calling the
Command.Main rather than spawning `rbmk <command>`).

So, because of all of this and because it seems cheap to
avoid this potential future issue, pave the way for a stdin
override inside of `rbmk` commands and subcommands.