BREAKING CHANGE(cliutils): ensure lazy help rendering
For ordinary commands, help rendering seems already lazy because
we explicitly need to invoke the Help method.

Conversely, for NewCommandWithSubCommands the help is passed
as a string, which means markdown rendering happens unconditionally.

This has negative implications including generation of garbled
output including terminal escape sequences when we put a process
in background using the shell.

To fix, make help rendering lazy also for NewCommandWithSubCommands.