feat: introduce file system abstraction
This abstraction is based on github.com/spf13/afero but adds
specific changes that make the code suitable for rbmk.

We will use this new functionality for testing of the rbmk tool
and possibly also to allow to virtualise the file system, a feature
that may be useful for embedding rbmk.

This change also required to extend the mocks to mock the new
filesystem types. The actual shared model between fsx and the mocks
package is a private model in the internal folder. The fsx package
exports all the private model definitions to fsx consumers.

While there upgrade dependencies.