refactor(fsx): RelativeFS => ContainedFS
This diff renames RelativeFS to ContainedFS, which better
reflects the dual semantics of relative paths and attempting
to jail the user paths within a base directory, with the
caveat that we cannot protect against symlink traversal though.

I have kept backward compatibility names for RelativeFS, so
this change should not break existing code.