feat(closepool): implement CloserFunc (#24)
This diff adds a CloserFunc type to closepool. We need this type to
seamlessly adapt closeables that do not implement `io.Closer` to the
`io.Closer` model. A concrete case where this would be needed is the
`quic.Conn` type, which has a different close pattern.

We will likely need this in
https://github.com/rbmk-project/dnscore/pull/18.